295 Move EcologyRecipe to common components
