+++ src/components/config.js
@@ -1,34 +1,34 @@
-import VAnimationDocs from "../../docs/victory-animation/docs.md";
-import VAreaDocs from "../../docs/victory-area/docs.md";
-import VAxisDocs from "../../docs/victory-axis/docs.md";
-import VBarDocs from "../../docs/victory-bar/docs.md";
-import VBrushContainerDocs from "../../docs/victory-brush-container/docs.md";
-import VCandlestickDocs from "../../docs/victory-candlestick/docs.md";
-import VChartDocs from "../../docs/victory-chart/docs.md";
-import VClipContainerDocs from "../../docs/victory-clip-container/docs.md";
-import VContainerDocs from "../../docs/victory-container/docs.md";
-import VErrorBarDocs from "../../docs/victory-errorbar/docs.md";
-import VGroupDocs from "../../docs/victory-group/docs.md";
-import VLabelDocs from "../../docs/victory-label/docs.md";
-import VLegendDocs from "../../docs/victory-legend/docs.md";
-import VLineDocs from "../../docs/victory-line/docs.md";
-import VPieDocs from "../../docs/victory-pie/docs.md";
-import VPortalDocs from "../../docs/victory-portal/docs.md";
-import VPrimitivesDocs from "../../docs/victory-primitives/docs.md";
-import VScatterDocs from "../../docs/victory-scatter/docs.md";
-import VSelectionContainerDocs from "../../docs/victory-selection-container/docs.md";
-import VSharedEventsDocs from "../../docs/victory-shared-events/docs.md";
-import VStackDocs from "../../docs/victory-stack/docs.md";
-import VThemeDocs from "../../docs/victory-theme/docs.md";
-import VTooltipDocs from "../../docs/victory-tooltip/docs.md";
-import VTransitionDocs from "../../docs/victory-transition/docs.md";
-import VVoronoiDocs from "../../docs/victory-voronoi/docs.md";
-import VVoronoiContainerDocs from "../../docs/victory-voronoi-container/docs.md";
-import VVoronoiTooltipDocs from "../../docs/victory-voronoi-tooltip/docs.md";
-import NativeTutorial from "../../docs/native/docs.md";
-import VZoomDocs from "../../docs/victory-zoom/docs.md";
-import VZoomContainerDocs from "../../docs/victory-zoom-container/docs.md";
-import CreateContainerDocs from "../../docs/create-container/docs.md";
+import VAnimationDocs from "../screens/docs/components/victory-animation/docs.md";
+import VAreaDocs from "../screens/docs/components/victory-area/docs.md";
+import VAxisDocs from "../screens/docs/components/victory-axis/docs.md";
+import VBarDocs from "../screens/docs/components/victory-bar/docs.md";
+import VBrushContainerDocs from "../screens/docs/components/victory-brush-container/docs.md";
+import VCandlestickDocs from "../screens/docs/components/victory-candlestick/docs.md";
+import VChartDocs from "../screens/docs/components/victory-chart/docs.md";
+import VClipContainerDocs from "../screens/docs/components/victory-clip-container/docs.md";
+import VContainerDocs from "../screens/docs/components/victory-container/docs.md";
+import VErrorBarDocs from "../screens/docs/components/victory-errorbar/docs.md";
+import VGroupDocs from "../screens/docs/components/victory-group/docs.md";
+import VLabelDocs from "../screens/docs/components/victory-label/docs.md";
+import VLegendDocs from "../screens/docs/components/victory-legend/docs.md";
+import VLineDocs from "../screens/docs/components/victory-line/docs.md";
+import VPieDocs from "../screens/docs/components/victory-pie/docs.md";
+import VPortalDocs from "../screens/docs/components/victory-portal/docs.md";
+import VPrimitivesDocs from "../screens/docs/components/victory-primitives/docs.md";
+import VScatterDocs from "../screens/docs/components/victory-scatter/docs.md";
+import VSelectionContainerDocs from "../screens/docs/components/victory-selection-container/docs.md";
+import VSharedEventsDocs from "../screens/docs/components/victory-shared-events/docs.md";
+import VStackDocs from "../screens/docs/components/victory-stack/docs.md";
+import VThemeDocs from "../screens/docs/components/victory-theme/docs.md";
+import VTooltipDocs from "../screens/docs/components/victory-tooltip/docs.md";
+import VTransitionDocs from "../screens/docs/components/victory-transition/docs.md";
+import VVoronoiDocs from "../screens/docs/components/victory-voronoi/docs.md";
+import VVoronoiContainerDocs from "../screens/docs/components/victory-voronoi-container/docs.md";
+import VVoronoiTooltipDocs from "../screens/docs/components/victory-voronoi-tooltip/docs.md";
+import NativeTutorial from "../screens/docs/components/native/docs.md";
+import VZoomDocs from "../screens/docs/components/victory-zoom/docs.md";
+import VZoomContainerDocs from "../screens/docs/components/victory-zoom-container/docs.md";
+import CreateContainerDocs from "../screens/docs/components/create-container/docs.md";
 
 export const config = [
   {
